Update help site documents and screenshots

+++ update.mts
@@ -113,26 +113,11 @@ async function saveLatestBuild() {
 }
 
 const helpLocales = [
-  {
-    docusaurus: "de",
-    crowdin: "de",
-  },
-  {
-    docusaurus: "es",
-    crowdin: "es-ES",
-  },
-  {
-    docusaurus: "fr",
-    crowdin: "fr",
-  },
-  {
-    docusaurus: "pt-BR",
-    crowdin: "pt-BR",
-  },
-  {
-    docusaurus: "id",
-    crowdin: "id",
-  },
+  { docusaurus: "de", crowdin: "de" },
+  { docusaurus: "es", crowdin: "es-ES" },
+  { docusaurus: "fr", crowdin: "fr" },
+  { docusaurus: "pt-BR", crowdin: "pt-BR" },
+  { docusaurus: "id", crowdin: "id" },
 ] as const;
 
 async function copyDirContentsRecursive(
@@ -154,7 +139,6 @@ async function copyDirContentsRecursive(
 
 async function copyFiles() {
   for (const locale of helpLocales) {
-    const sourceDocsDir = `${projectRoot}/docs`;
     const localizedGuidesDir = `${projectRoot}/translations/${locale.crowdin}/Guides`;
     const dest = `${projectRoot}/i18n/${locale.docusaurus}/`;
     const docusaurusDest = `${dest}docusaurus-plugin-content-docs`;
@@ -164,9 +148,6 @@ async function copyFiles() {
     console.log(`Creating directory ${currentDest}`);
     await Deno.mkdir(docusaurusDest, { recursive: true });
     await Deno.mkdir(currentDest, { recursive: true });
-
-    console.log(`Copying files from ${sourceDocsDir} to ${currentDest}`);
-    await copyDirContentsRecursive(sourceDocsDir, `${currentDest}`);
 
     console.log(`Copying files from ${localizedGuidesDir} to ${dest}`);
     await copyDirContentsRecursive(localizedGuidesDir, dest);
@@ -183,36 +164,24 @@ async function deleteExistingFiles() {
   console.log(`Deleting most files in ${docsDir}`);
   // delete all files except 	docs/getting-started.json and docs/getting-started.mdx
 
-  const filesToKeep = ["getting-started.json", "getting-started.mdx"];
+  // TODO Since we're no longer fetching from Notion, we can't delete the files in the docs directory
+  // Consider updating this to use Crowdin as the source of truth for the docs
+  const deleteEnglishFiles = false;
+  if (deleteEnglishFiles) {
+    const filesToKeep = ["getting-started.json", "getting-started.mdx"];
 
-  for await (const dirEntry of Deno.readDir(docsDir)) {
-    if (dirEntry.isFile && !filesToKeep.includes(dirEntry.name)) {
-      await Deno.remove(`${docsDir}/${dirEntry.name}`);
+    for await (const dirEntry of Deno.readDir(docsDir)) {
+      if (dirEntry.isFile && !filesToKeep.includes(dirEntry.name)) {
+        await Deno.remove(`${docsDir}/${dirEntry.name}`);
+      }
     }
   }
 
-  const i18nDir = `${projectRoot}/i18n`;
-  console.log(`Deleting ${i18nDir}`);
-  await Deno.remove(i18nDir, { recursive: true });
-}
-
-async function fetchNotionDocs() {
-  const child = new Deno.Command("bash", {
-    args: ["pull_docs.sh"],
-    stdout: "piped",
-    stderr: "piped",
-  }).spawn();
-
-  copy(readerFromStreamReader(child.stdout.getReader()), Deno.stdout);
-  copy(readerFromStreamReader(child.stderr.getReader()), Deno.stderr);
-
-  const status = await child.status;
-
-  if (!status.success) {
-    const code = status.code;
-    throw new Error(
-      `Failed to fetch Notion docs. pull_docs.sh exited with code ${code}`
-    );
+  // Recursively delete all files in the i18n directory, except for PNG files
+  for await (const dirEntry of walk(`${projectRoot}/i18n`, { maxDepth: 1 })) {
+    if (dirEntry.isFile && !dirEntry.name.endsWith(".png")) {
+      await Deno.remove(dirEntry.path);
+    }
   }
 }
 
@@ -342,10 +311,6 @@ async function runChecks() {
 try {
   console.log("--- Deleting existing files ---");
   await deleteExistingFiles();
-  console.log();
-
-  console.log("--- Fetching latest docs from Notion ---");
-  await fetchNotionDocs();
   console.log();
 
   console.log("--- Fetching latest translations from Crowdin ---");
